+++ src/libs/nav2_util/include/nav2_util/sensors/amcl_laser.h
@@ -29,7 +29,7 @@
 #ifndef AMCL_LASER_H
 #define AMCL_LASER_H
 
-#include "amcl_sensor.h"
+#include "nav2_util/sensors/amcl_sensor.h"
 #include "nav2_util/map.h"
 
 namespace amcl
