+++ src/libs/nav2_util/include/nav2_util/sensors/amcl_odom.h
@@ -29,7 +29,7 @@
 #ifndef AMCL_ODOM_H
 #define AMCL_ODOM_H
 
-#include "amcl_sensor.h"
+#include "nav2_util/sensors/amcl_sensor.h"
 #include "nav2_util/pf/pf_pdf.h"
 
 namespace amcl
