+++ src/libs/nav2_util/src/sensors/amcl_odom.cpp
@@ -32,7 +32,7 @@
 #include <sys/types.h> // required by Darwin
 #include <math.h>
 
-#include "amcl/sensors/amcl_odom.h"
+#include "nav2_util/sensors/amcl_odom.h"
 
 using namespace amcl;
