+++ src/nav2_localization/amcl/package.xml
@@ -21,6 +21,7 @@
   <maintainer email="davidvlu@gmail.com">David V. Lu!!</maintainer>
   <maintainer email="mfergs7@gmail.com">Michael Ferguson</maintainer>
   <maintainer email="ahoy@fetchrobotics.com">Aaron Hoy</maintainer>
+  <maintainer email="mohammad.haghighipanah@intel.com">Mohammad Haghighipanah</maintainer>
   <license>LGPL</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
